Fix running test_qrspec when building out-of-source
It is a common practice for CMake builds (and less common but still
valid for autotools) to build out-of-source, that is use a separate
directory for built executables than for sources.  However, when that is
used, test_qrspec is unable to find "frame" and crashes.  Fix it by
passing the source directory as compile-time definition, and therefore
making test_qrspec use the correct path.