cmake: Always build libdecoder as static library
libdecoder is used only for tests, and implicitly relies
on printBinary() function that is not included in the library.  This
causes it to fail to build as a shared library on systems strict about
symbols (Darwin or e.g. Linux with -Wl,-z,defs).  Force static library
to avoid those problems.