writeANSI: use strcpy instead of strncat
strncat(3) needs to traverse through the buffer to append the string
into the end of old string. This will have a big performance penalty
when we have a very large input.

Replace it with strcpy(3) and a pointer to track the end of current
buffer.

Accidently, this change also silence gcc's -Wstringop-overflow warning,
since gcc thinks our issue to strncat(3) to stop right at the NULL
terminator maybe an error.