writeANSI: allocate buffer based on longer color string
On a modern compiler, the compiler should be able to figure out both
black_s and white_s are compile time constant and fold this change to
constant. And we can easier argue why we need to allow that much of
memory.

This change also help if we decide to support another kind of weird
terminal color scheme in future.